# src/configs/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "InsecureRestAPI",
    "version": "1.0",
    "description": "<b>InsecureRestAPI - An insecure NodeJS/Express/MongoDB REST API for educational purposes.</b>\n\nYou can select a development or production server to test the API. Most operations require authentication via a user specific JWT token. To retrieve a JWT token for a user you can use the Site <code>'/api/v1/site/sign-in'</code> operation below and then copy the value of the <b>accessToken</b> field. This value can then be entered when you click on the <b>Authorize</b> button or lock icons.\n\n[View Schema](/docs/openapi.json)",
    "license": {
      "name": "GPLv3",
      "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
    },
    "contact": {
      "name": "Kevin A. Lee (kadraman)",
      "url": "https://github.com/kadraman"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "InsecureRestAPI - development "
    }
  ],
  "paths": {
...
    "/api/v1/messages": {
      "get": {
        "tags": [
          "Messages"
        ],
        "summary": "Find messages by keyword(s)",
        "description": "Gets all existing messages searching by %keyword% format",
        "operationId": "getMessages",
        "parameters": [
          {
            "name": "keywords",
            "in": "query",
            "description": "Keyword(s) search for messages to be found.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Offset of the starting record. 0 indicates the first record.",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum records to return. The maximum value allowed is 50.",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/success"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/success"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Messages"
        ],
        "summary": "Create new message",
        "description": "Creates a new message",
        "operationId": "createMessage",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/success"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/success"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/message"
              }
            }
          }
        }
      }
    },
    "/api/v1/messages/{id}": {
      "get": {
        "tags": [
          "Messages"
        ],
        "summary": "Get a message",
        "description": "Gets an existing message",
        "operationId": "getMessageById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Id of the message to be retrieved. Cannot be empty."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/success"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/success"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "404": {
            "description": "Message Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Messages"
        ],
        "summary": "Update a message",
        "description": "Updates an existing message",
        "operationId": "updateMessage",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Id of the message to be updated. Cannot be empty."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/success"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/success"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "404": {
            "description": "Message Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Messages"
        ],
        "summary": "Delete a message",
        "description": "Deletes an existing message",
        "operationId": "deleteMessage",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Id of the message to be deleted. Cannot be empty."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/success"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/success"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "404": {
            "description": "Message Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/failure"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "message": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string",
            "example": "87cd9f67-d160-4487-ba95-cb889a47cf2d"
          },
          "text": {
            "type": "string",
            "example": "This is an example message"
          },
          "sent_date": {
            "type": "string",
            "example": "YYYY-MM-DDTHH:mm:ssZ"
          },
          "read_date": {
            "type": "string",
            "example": "YYYY-MM-DDTHH:mm:ssZ"
          },
          "is_read": {
            "type": "boolean",
            "example": true
          },
          "is_deleted": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "user_id",
          "text",
          "sent_date",
          "read_date",
          "is_read",
          "is_deleted"
        ],
        "xml": {
          "name": "message"
        }
      },
      "success": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "message": {
            "type": "string",
            "example": "string"
          },
          "timestamp": {
            "type": "string",
            "example": "YYYY-MM-DDTHH:mm:ssZ"
          },
          "data": {
            "type": "object",
            "properties": {}
          }
        },
        "required": [
          "status",
          "message",
          "timestamp",
          "data"
        ],
        "xml": {
          "name": "success"
        }
      },
      "failure": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "failure"
          },
          "message": {
            "type": "string",
            "example": "string"
          },
          "timestamp": {
            "type": "string",
            "example": "YYYY-MM-DDTHH:mm:ssZ"
          },
          "data": {
            "type": "object",
            "properties": {}
          }
        },
        "required": [
          "status",
          "message",
          "timestamp",
          "data"
        ],
        "xml": {
          "name": "failure"
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Access Token"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}